Follow modern _ naming
With backwards compatibility for uri-redis
