+++ lib/uri-redis.rb
@@ -1,1 +1,5 @@
-require "uri-redis/redis"
+# frozen_string_literal: true
+
+# Compatibility shim for uri-redis.rb
+# This file maintains backward compatibility for code that requires 'uri-redis'
+require_relative 'uri_redis'
