+++ lib/uri-valkey.rb
@@ -1,1 +1,5 @@
-require "uri-valkey/valkey"
+# frozen_string_literal: true
+
+# Compatibility shim for uri-valkey.rb
+# This file maintains backward compatibility for code that requires 'uri-valkey'
+require_relative 'uri_valkey'
