+++ lib/uri_redis.rb
@@ -1,0 +1,1 @@
+require_relative 'uri_redis/redis'
