+++ lib/uri_redis/redis.rb
@@ -1,21 +1,23 @@
-require "uri"
-require "uri/generic"
+# lib/uri-redis/redis.rb
 
-# URI::Valkey - adds support for Valkey URIs to core.
+require 'uri'
+require 'uri/generic'
+
+# URI::Redis - adds support for Redis URIs to core.
 module URI
-  # Valkey URI
+  # Redis URI
   #
   # This is a subclass of URI::Generic and supports the following URI formats:
   #
-  #   valkey://host:port/dbindex
+  #   redis://host:port/dbindex
   #
   # @example
-  #   uri = URI::Valkey.build(host: "localhost", port: 6379, db: 2, key: "v1:arbitrary:key")
-  #   uri.to_s #=> "valkey://localhost:6379/2/v1:arbitrary:key"
+  #   uri = URI::Redis.build(host: "localhost", port: 6379, db: 2, key: "v1:arbitrary:key")
+  #   uri.to_s #=> "redis://localhost:6379/2/v1:arbitrary:key"
   #
-  #   uri = URI::Valkey.build(host: "localhost", port: 6379, db: 2)
-  #   uri.to_s #=> "valkey://localhost:6379/2"
-  class Valkey < URI::Generic
+  #   uri = URI::Redis.build(host: "localhost", port: 6379, db: 2)
+  #   uri.to_s #=> "redis://localhost:6379/2"
+  class Redis < URI::Generic
     DEFAULT_PORT = 6379
     DEFAULT_DB = 0
 
@@ -32,16 +34,16 @@ module URI
       return if path.nil?
 
       self.path ||= "/#{DEFAULT_DB}"
-      (self.path.split("/")[2..] || []).join("/")
+      (self.path.split('/')[2..] || []).join('/')
     end
 
     def key=(val)
-      self.path = "/" << [db, val].join("/")
+      self.path = '/' << [db, val].join('/')
     end
 
     def db
       self.path ||= "/#{DEFAULT_DB}"
-      (self.path.split("/")[1] || DEFAULT_DB).to_i
+      (self.path.split('/')[1] || DEFAULT_DB).to_i
     end
 
     def db=(val)
@@ -51,13 +53,13 @@ module URI
       self.path
     end
 
-    # Returns a hash suitable for sending to Valkey.new.
+    # Returns a hash suitable for sending to Redis.new.
     # The hash is generated from the host, port, db and
     # password from the URI as well as any query vars.
     #
     # e.g.
     #
-    #      uri = URI.parse "valkey://127.0.0.1/6/?timeout=5"
+    #      uri = URI.parse "redis://127.0.0.1/6/?timeout=5"
     #      uri.conf
     #        # => {:db=>6, :timeout=>"5", :host=>"127.0.0.1", :port=>6379}
     #
@@ -66,7 +68,7 @@ module URI
         host: host,
         port: port,
         db: db,
-        ssl: scheme == 'valkeys'
+        ssl: scheme == 'rediss'
       }.merge(parse_query(query))
       hsh[:password] = password if password
       hsh[:timeout] = hsh[:timeout].to_i if hsh.key?(:timeout)
@@ -82,10 +84,10 @@ module URI
     # Based on: https://github.com/chneukirchen/rack/blob/master/lib/rack/utils.rb
     # which was originally based on Mongrel
     def parse_query(query, delim = nil)
-      delim ||= "&;"
+      delim ||= '&;'
       params = {}
-      (query || "").split(/[#{delim}] */n).each do |p|
-        k, v = p.split("=", 2).map { |str| str } # NOTE: uri_unescape
+      (query || '').split(/[#{delim}] */n).each do |p|
+        k, v = p.split('=', 2).map { |str| str } # NOTE: uri_unescape
         k = k.to_sym
         if (cur = params[k])
           if cur.instance_of?(Array)
@@ -102,10 +104,28 @@ module URI
   end
 
   if URI.respond_to?(:register_scheme)
-    URI.register_scheme 'VALKEY', Valkey
-    URI.register_scheme 'VALKEYS', Valkey
+    URI.register_scheme 'REDIS', Redis
+    URI.register_scheme 'REDISS', Redis
   else
-    @@schemes['VALKEY'] = Valkey
-    @@schemes['VALKEYS'] = Valkey
+    @@schemes['REDIS'] = Redis
+    @@schemes['REDISS'] = Redis
+  end
+end
+
+if defined?(Redis)
+  # Usage:
+  # using RedisURIRefinement
+  # redis = Redis.new
+  # redis.uri
+  module RedisURIRefinement
+    refine Redis do
+      def uri
+        URI.parse @client.id
+      end
+
+      def self.uri(conf = {})
+        URI.parse format('%s://%s:%s/%s', conf[:ssl] ? 'rediss' : 'redis', conf[:host], conf[:port], conf[:db])
+      end
+    end
   end
 end
