+++ lib/uri_redis/version.rb
