+++ lib/uri_valkey.rb
@@ -1,0 +1,1 @@
+require_relative 'uri_valkey/valkey'
