+++ lib/uri_valkey/valkey.rb
@@ -1,24 +1,21 @@
-# lib/uri-redis/redis.rb
-
 require 'uri'
 require 'uri/generic'
 
-
-# URI::Redis - adds support for Redis URIs to core.
+# URI::Valkey - adds support for Valkey URIs to core.
 module URI
-  # Redis URI
+  # Valkey URI
   #
   # This is a subclass of URI::Generic and supports the following URI formats:
   #
-  #   redis://host:port/dbindex
+  #   valkey://host:port/dbindex
   #
   # @example
-  #   uri = URI::Redis.build(host: "localhost", port: 6379, db: 2, key: "v1:arbitrary:key")
-  #   uri.to_s #=> "redis://localhost:6379/2/v1:arbitrary:key"
+  #   uri = URI::Valkey.build(host: "localhost", port: 6379, db: 2, key: "v1:arbitrary:key")
+  #   uri.to_s #=> "valkey://localhost:6379/2/v1:arbitrary:key"
   #
-  #   uri = URI::Redis.build(host: "localhost", port: 6379, db: 2)
-  #   uri.to_s #=> "redis://localhost:6379/2"
-  class Redis < URI::Generic
+  #   uri = URI::Valkey.build(host: "localhost", port: 6379, db: 2)
+  #   uri.to_s #=> "valkey://localhost:6379/2"
+  class Valkey < URI::Generic
     DEFAULT_PORT = 6379
     DEFAULT_DB = 0
 
@@ -54,13 +51,13 @@ module URI
       self.path
     end
 
-    # Returns a hash suitable for sending to Redis.new.
+    # Returns a hash suitable for sending to Valkey.new.
     # The hash is generated from the host, port, db and
     # password from the URI as well as any query vars.
     #
     # e.g.
     #
-    #      uri = URI.parse "redis://127.0.0.1/6/?timeout=5"
+    #      uri = URI.parse "valkey://127.0.0.1/6/?timeout=5"
     #      uri.conf
     #        # => {:db=>6, :timeout=>"5", :host=>"127.0.0.1", :port=>6379}
     #
@@ -69,7 +66,7 @@ module URI
         host: host,
         port: port,
         db: db,
-        ssl: scheme == 'rediss'
+        ssl: scheme == 'valkeys'
       }.merge(parse_query(query))
       hsh[:password] = password if password
       hsh[:timeout] = hsh[:timeout].to_i if hsh.key?(:timeout)
@@ -105,21 +102,10 @@ module URI
   end
 
   if URI.respond_to?(:register_scheme)
-    URI.register_scheme 'REDIS', Redis
-    URI.register_scheme 'REDISS', Redis
+    URI.register_scheme 'VALKEY', Valkey
+    URI.register_scheme 'VALKEYS', Valkey
   else
-    @@schemes['REDIS'] = Redis
-    @@schemes['REDISS'] = Redis
-  end
-end
-
-# Adds a URI method to Redis
-class Redis
-  def self.uri(conf = {})
-    URI.parse format('%s://%s:%s/%s', conf[:ssl] ? 'rediss' : 'redis', conf[:host], conf[:port], conf[:db])
-  end
-
-  def uri
-    URI.parse @client.id
+    @@schemes['VALKEY'] = Valkey
+    @@schemes['VALKEYS'] = Valkey
   end
 end
